chore(Makefile): enable -Werror, -Wall and -Wextra flags chore(so_long.h): change x and y player coordinates to int32_t chore(so_long.h): change player health, wallet and keys to uint32_t chore(so_long.h): remove unused player_idle and player_walk arrays refactor(files.c): remove unused function argument refactor(files.c): remove commented out code refactor(files.c): change load_texture_files function argument to char pointer refactor(hooks.c): change mlx_close_window argument to void pointer refactor(hooks.c): change g_mlx->width to (unsigned int)g_mlx->width refactor(player_move.c): change g_mlx->height to (int32_t)g_mlx->height refactor(player_move.c): change g_img->img->height to (int32_t)g_img->img->height

+++ src/player_move.c
@@ -6,7 +6,7 @@
 /*   By: jensbouma <jensbouma@student.codam.nl>       +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/05/12 17:23:15 by jensbouma     #+#    #+#                 */
-/*   Updated: 2023/05/15 12:12:58 by jensbouma     ########   odam.nl         */
+/*   Updated: 2023/05/16 00:12:35 by jensbouma     ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,13 +42,13 @@ void	player_jump(void)
 void	player_fall(void)
 {
 	if (g_player->jump_state == FALL
-		&& g_player->y < g_mlx->height - g_img->img->height)
+		&& g_player->y < g_mlx->height - (int32_t)g_img->img->height)
 	{
-		if (g_player->y + g_img->img->height + 5 < g_mlx->height)
+		if (g_player->y + (int32_t)g_img->img->height + 5 < g_mlx->height)
 			g_player->y += 5;
 		else
 		{
-			g_player->y = g_mlx->height - g_img->img->height;
+			g_player->y = g_mlx->height - (int32_t)g_img->img->height;
 			g_player->jump_state = STAND;
 		}
 	}
